Did minor fix as you advised
