+++ models/contactMongoose.js
@@ -1,0 +1,37 @@
+const { Schema, model } = require("mongoose");
+const { handleMongooseError } = require("../helpers");
+
+const phoneRegexp = /^(?:\(\d{3}\)|\d{3})[- ]?\d{3}[- ]?\d{4}$/;
+const emailRegexp = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+
+const contactSchema = new Schema(
+  {
+    name: {
+      type: String,
+      required: true,
+    },
+    email: {
+      type: String,
+      match: emailRegexp,
+      required: true,
+    },
+    phone: {
+      type: String,
+      match: phoneRegexp,
+      required: true,
+    },
+    favorite: {
+      type: Boolean,
+      default: false,
+    },
+  },
+  { versionKey: false, timestamps: true }
+);
+
+contactSchema.post("save", handleMongooseError);
+
+const Contact = model("contact", contactSchema);
+
+module.exports = {
+  Contact,
+};
